begin reorganize models and loglistener
